Fix compilation without X86 or GFXBOARD support
On Apple Silicon, WITH_X86 was correctly NOT defined but some of the code did not compile correctly with this turned off.

Also, GFXBOARD support can’t compile without WITH_X86 so this is now correctly handled too and the code now also compiles correctly when GFXBOARD support is off.

configure.ac was also updated to autoconf version 2.72.